Refactor createLLVMMod to use absl and move to a different file
